changes based on comments #4608
