+++ src/Service/DashboardTabPanel/TabBodyLoader/DossiersAdresseEmailAverifierTabBodyLoader.php
@@ -29,12 +29,6 @@ class DossiersAdresseEmailAverifierTabBodyLoader extends AbstractTabBodyLoader
             'sortBy' => $this->tabQueryParameters->sortBy ?? 'createdAt',
             'direction' => $this->tabQueryParameters->orderBy ?? 'DESC',
         ];
-        if ($this->tabQueryParameters->partners && \count($this->tabQueryParameters->partners) > 0) {
-            $filters['partenaires'] = $this->tabQueryParameters->partners;
-        }
-        if (null !== $this->tabQueryParameters->queryCommune && '' !== $this->tabQueryParameters->queryCommune) {
-            $filters['communes[]'] = $this->tabQueryParameters->queryCommune;
-        }
         $tabBody->setFilters($filters);
         $tabBody->setTemplate('back/dashboard/tabs/dossiers_a_verifier/_body_dossier_adresse_email_a_verifier.html.twig');
     }
